break out astrometry and kinematics tests
